src/requirements.txt
Django==1.8.13
django-braces==1.9.0
django-crispy-forms==1.6.0
django-filter==0.13.0
django-registration-redux==1.4
django-tables2==1.2.3
Pillow==3.3.0
